Turn abstract base class into an interface
