+++ src/index.ts
@@ -2,8 +2,7 @@ import SegfaultHandler from 'segfault-handler';
 
 export * from './launchpads/MK2/LaunchpadMK2';
 export * from './launchpads/MK3/LaunchpadMK3';
-export * from './launchpads/base/BaseLaunchpad';
-export * from './launchpads/base/types';
+export * from './launchpads/base/ILaunchpad';
 export * from './launchpads/autoDetect';
 export * from './launchpads/autoDetect';
 export * as colors from './colorHelpers';
