+++ test/mocking-midi.ts
@@ -8,10 +8,7 @@ import * as midi from 'midi';
 const mockedInput = jest.mocked(new midi.Input(), true);
 const mockedOutput = jest.mocked(new midi.Output(), true);
 
-jest.mocked(midi.Input).mockImplementation(() => {
-  console.log('xyz');
-  return mockedInput;
-});
+jest.mocked(midi.Input).mockReturnValue(mockedInput);
 jest.mocked(midi.Output).mockReturnValue(mockedOutput);
 
 /**
